+++ src/aks-preview/HISTORY.rst
@@ -10,7 +10,30 @@ If there is no rush to release a new version, please just add a description of t
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
 
 Pending
+++++++
+
+0.5.145
 +++++++
+* Add support for option --aks-custom-headers to some aks commands
+  * aks get-credentials
+  * aks nodepool scale
+  * aks nodepool update
+  * aks enable-addons
+  * aks show
+  * aks scale
+
+0.5.144
++++++++
+* Fix setup network profile with network observability due to incorrect property
+
+0.5.143
++++++++
+* Vendor new SDK and bump API version to 2023-05-02-preview.
+* Add `--enable-network-observability` flag to `az aks create` and `az aks update`.
+
+0.5.142
++++++++
+* Deprecate option names `--enable-azuremonitormetrics` and `--disable-azuremonitormetrics`, use `--enable-azure-monitor-metrics` and `--disable-azure-monitor-metrics` instead, so as to be consistent with the option names in official azure-cli. Fix issue `\#26600 <https://github.com/Azure/azure-cli/issues/26600>`_.
 
 0.5.141
 +++++++